Add UpdatableEditableAdapter and use it in ScheduleRecyclerViewAdapter.kt
