+++ app/src/main/java/com/justsoft/nulpschedule/ui/recyclerview/SwipeAndDragCallback.kt
@@ -2,13 +2,16 @@ package com.justsoft.nulpschedule.ui.recyclerview
 
 import android.content.Context
 import android.content.res.ColorStateList
-import android.graphics.*
+import android.graphics.Canvas
+import android.graphics.Paint
+import android.graphics.PorterDuff
+import android.graphics.PorterDuffXfermode
 import android.graphics.drawable.GradientDrawable
+import android.util.Log
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
 import com.justsoft.nulpschedule.R
-import com.justsoft.nulpschedule.fragments.scheduleselectfragment.ScheduleRecyclerViewAdapter
 
 class SwipeAndDragCallback(
     context: Context
@@ -45,10 +48,17 @@ class SwipeAndDragCallback(
         viewHolder: RecyclerView.ViewHolder,
         target: RecyclerView.ViewHolder
     ): Boolean {
-        (recyclerView.adapter as ScheduleRecyclerViewAdapter).move(
-            viewHolder.adapterPosition,
-            target.adapterPosition
-        )
+        if (recyclerView.adapter !is UpdatableEditableAdapter<*, *>) {
+            Log.w(
+                "SwipeAndDragCallback",
+                "Supplied RecyclerView's adapter does not extend UpdatableEditableAdapter!"
+            )
+        } else {
+            (recyclerView.adapter as UpdatableEditableAdapter<*, *>).moveItem(
+                viewHolder.adapterPosition,
+                target.adapterPosition
+            )
+        }
         onMoveCallback.onMove(viewHolder, target)
         return true
     }
